pyproject.toml
[tool.poetry]
name = "work-atlas"
version = "0.1.0"
description = "Multi-tenant employee map backend"
authors = ["Acme Corp <dev@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.15"
django-tenants = "^3.6"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
django-environ = "^0.11"
psycopg2-binary = "^2.9"
Pillow = "^10.0"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2"
pytest = "^8.0"
pytest-django = "^4.8"
factory_boy = "^3.3"
freezegun = "^1.4"
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"
pre-commit = "^3.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

